Italic validation (#80)
* Add rule to detect italic

* Extract file helper functions

* Remove italic formattings

* Report line number for rules validation errors

* Improve validation error handling

# Conflicts:
#	tests/tutorial.js
