+++ content/tutorials/portenta-h7/por-ard-usb/content.md
@@ -6,7 +6,7 @@ It is possible to configure the Portenta H7 to act as a USB host in a way that a
 ### You Will Learn
 
 -   How to configure the Portenta H7 as a USB host
--   How to use the *KeyboardController* library to establish a USB connection with the Portenta H7
+-   How to use the **KeyboardController** library to establish a USB connection with the Portenta H7
 -   To write a small program that reads button presses from a keyboard
 -   How to debug the data being sent from the peripheral to the Portenta H7 using an additional Arduino board.
 
@@ -45,10 +45,10 @@ To achieve this you will use the [USBHOST](https://os.mbed.com/handbook/USBHost)
 
 Thanks to USB OTG (On The Go) specification the Portenta H7 can switch between host and client modes automatically. The code that enables USB OTG in Portenta H7 is included in the **mbed** core and it is ready to function without additional configurations. Some related classes to make use of the USB port for different use cases are:
 
-* `USBCDC_ECM` > Ethernet over USB (See [here](https://os.mbed.com/docs/mbed-os/v5.15/apis/usbcdc-ecm.html))
-* `USBMSD` > Mass storage device over USB (See [here](https://os.mbed.com/docs/mbed-os/v5.15/apis/usbmsd.html))
-* `USBCDC` > Serial port over USB (See [here](https://os.mbed.com/docs/mbed-os/v5.15/apis/usbcdc.html))
-* `USBHID` > USB human interface devices e.g. mouse or keyboard (See [here](https://os.mbed.com/docs/mbed-os/v5.15/apis/usbhid.html))
+- `USBCDC_ECM` > Ethernet over USB (See [here](https://os.mbed.com/docs/mbed-os/v5.15/apis/usbcdc-ecm.html))
+- `USBMSD` > Mass storage device over USB (See [here](https://os.mbed.com/docs/mbed-os/v5.15/apis/usbmsd.html))
+- `USBCDC` > Serial port over USB (See [here](https://os.mbed.com/docs/mbed-os/v5.15/apis/usbcdc.html))
+- `USBHID` > USB human interface devices e.g. mouse or keyboard (See [here](https://os.mbed.com/docs/mbed-os/v5.15/apis/usbhid.html))
 
 ### 1. The Basic Setup
 
@@ -60,7 +60,7 @@ Begin by plugging in your Portenta board to the computer using a USB-C  cable an
 
 Let's create a sketch that handles the USB connections and  modifies the state of the LEDs with each press on the R, G or B keys. When you press the key first time the LEDs will switch on and to turn them off, you will need to press the corresponding key again.
 
-As programming the USB protocol that allows the board to handle USB devices is an arduous task, you will use a pre-built example called *KeyboardController*. To find it make sure you have the board description for the Arduino Portenta H7 installed using the Boards Manager.
+As programming the USB protocol that allows the board to handle USB devices is an arduous task, you will use a pre-built example called **KeyboardController**. To find it make sure you have the board description for the Arduino Portenta H7 installed using the Boards Manager.
 
 ![Select the Arduino Portenta H7 (M7 core) in the board selector](assets/por_ard_usbh_select_board.png)
