+++ tests/validation-error.js
@@ -1,0 +1,9 @@
+var ValidationError = class ValidationError {
+    constructor(message, file, linenumber){
+        this.message = message;
+        this.file = file;
+        this.linenumber = linenumber;
+    }
+}
+
+module.exports = { ValidationError }
